Move GetWorkloadPackRecords into partial class file of NetSdkMsiInstallerClient
